+++ opts.h
@@ -81,7 +81,7 @@ typedef struct opts {
 #ifdef HAVE_LOCAL_PROCINFO
 	unsigned int lprocinfo : 1;
 #endif /* HAVE_LOCAL_PROCINFO */
-	unsigned int certgen_writeall: 1;
+	unsigned int certgen_writeall : 1;
 #ifndef OPENSSL_NO_ENGINE
 	char *openssl_engine;
 #endif /* !OPENSSL_NO_ENGINE */